Gravity first pass complete

+++ particle.h
@@ -7,8 +7,8 @@ class Particle
 {
     private:
     public:
-        int x = 0;
-        int y = 0;
+        double long x = 0;
+        double long y = 0;
         long double mass = 0.0;
         long double x_vel = 0.0;
         long double y_vel = 0.0;
